pyproject.toml
[project]
name = "reseller-tool"
version = "0.1.0"
description = "eBay vs AliExpress product research and margin analysis tool for dropshippers"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Tepfambiz" }]

dependencies = [
    "streamlit>=1.40.0",
    "pandas>=2.2.0",
    "pytrends>=4.9.0",
    "google-search-results>=2.4.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "python-Levenshtein>=0.25.0",
    "plotly>=5.24.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
reseller = "reseller_tool.cli:main"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
